src/test/java/de/gematik/demis/reportprocessingservice/security/converters/JwtAuthoritiesConverterTest.java
/*
 * Copyright [2024], gematik GmbH
 *
 * Licensed under the EUPL, Version 1.2 or - as soon they will be approved by the
 * European Commission – subsequent versions of the EUPL (the "Licence").
 * You may not use this work except in compliance with the Licence.
 *
 * You find a copy of the Licence in the "Licence" file or at
 * https://joinup.ec.europa.eu/collection/eupl/eupl-text-eupl-12
 *
 * Unless required by applicable law or agreed to in writing,
 * software distributed under the Licence is distributed on an "AS IS" basis,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either expressed or implied.
 * In case of changes by gematik find details in the "Readme" file.
 *
 * See the Licence for the specific language governing permissions and limitations under the Licence.
 */

package de.gematik.demis.reportprocessingservice.security.converters;

import static org.assertj.core.api.Assertions.assertThat;
import static org.mockito.ArgumentMatchers.anyString;
import static org.mockito.Mockito.when;

import com.fasterxml.jackson.core.JsonProcessingException;
import de.gematik.demis.reportprocessingservice.properties.OAuth2Properties;
import java.util.Collection;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.Mock;
import org.mockito.junit.jupiter.MockitoExtension;
import org.springframework.security.core.GrantedAuthority;
import org.springframework.security.oauth2.jwt.Jwt;

@ExtendWith(MockitoExtension.class)
class JwtAuthoritiesConverterTest {
  @Mock private OAuth2Properties auth2Properties;
  @Mock private Jwt jwt;

  @Test
  void givenInvalidJSONWhenConvertThenNoAuthorities() throws JsonProcessingException {
    when(jwt.getClaimAsMap(anyString())).thenReturn(null);
    JwtAuthoritiesConverter converter = new JwtAuthoritiesConverter(auth2Properties);

    final Collection<GrantedAuthority> authorities = converter.convert(jwt);
    assertThat(authorities).isEmpty();
  }
}
